requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1 